Address review suggestions & post-rebase fixes
