+++ cardano-node/app/chairman.hs
@@ -17,7 +17,7 @@ import           Ouroboros.Consensus.Protocol.Abstract (SecurityParam (..))
 import           Ouroboros.Consensus.Node.ProtocolInfo.Abstract (NumCoreNodes (..))
 import           Ouroboros.Consensus.NodeId (CoreNodeId)
 
-import           Cardano.Common.CommonCLI (CommonCLI, mergeConfiguration, parseCommonCLI)
+import           Cardano.Common.CommonCLI
 import           Cardano.Common.Protocol (Protocol, SomeProtocol(..), fromProtocol)
 import           Cardano.Node.Parsers (parseCoreNodeId, parseProtocol)
