+++ cardano-node/src/Cardano/Node/Types.hs
@@ -1,0 +1,22 @@
+{-# LANGUAGE RankNTypes           #-}
+
+module Cardano.Node.Types
+  ( Protocol(..)
+  , parseProtocol
+  , parseProtocolAsCommand
+  )
+where
+
+import Prelude
+import Data.Foldable
+import Options.Applicative
+
+data Protocol
+  = ByronLegacy
+  | BFT
+  | Praos
+  | MockPBFT
+  | RealPBFT
+  deriving (Show)
+
+
